Rename old search and widget pages
