+++ libs/common/src/lib/search-widget/search/deprecated-search.component.scss
@@ -1,4 +1,4 @@
-@import '../../../../../../apps/dashboard/src/variables';
+@import 'apps/dashboard/src/variables';
 
 .search-page {
   .search-bar-section {
