+++ libs/common/src/lib/search-widget/search/index.ts
@@ -1,2 +1,2 @@
-export * from './search.component';
+export * from './deprecated-search.component';
 export * from './search.module';
