+++ libs/common/src/lib/search-widget/search/search.module.ts
@@ -1,12 +1,12 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { SearchComponent } from './search.component';
+import { DeprecatedSearchComponent } from './deprecated-search.component';
 import { PaIconModule, PaTogglesModule } from '@guillotinaweb/pastanaga-angular';
 import { TranslateModule } from '@ngx-translate/core';
 
 @NgModule({
   imports: [CommonModule, PaTogglesModule, TranslateModule, PaIconModule],
-  declarations: [SearchComponent],
-  exports: [SearchComponent],
+  declarations: [DeprecatedSearchComponent],
+  exports: [DeprecatedSearchComponent],
 })
 export class SearchModule {}
