+++ libs/common/src/lib/search-widget/widgets/deprecated-widget-generator.component.scss
@@ -1,4 +1,4 @@
-@import '../../../../../../apps/dashboard/src/variables';
+@import 'apps/dashboard/src/variables';
 
 .widget-generator {
   $margin-bottom-title: rhythm(4);
