+++ libs/common/src/lib/search-widget/widgets/deprecated-widget-generator.models.ts
