+++ libs/common/src/lib/search-widget/widgets/index.ts
@@ -1,5 +1,5 @@
 export * from './copilot/copilot-modal.component';
-export * from './widget-features';
-export * from './widget-generator.component';
-export * from './widget-generator.models';
-export * from './widgets.module';
+export * from './deprecated-widget-features';
+export * from './deprecated-widget-generator.component';
+export * from './deprecated-widget-generator.models';
+export * from './deprecated-widgets.module';
